+++ Server/main.c
@@ -19,7 +19,8 @@
 #include "./include/sockets/handleClient.h"
 #include "./include/sockets/socketFunctions.h"
 #include "./include/sockets/handleUpdateClients.h"
-#include "include/utils/handleJson.h"
+#include "./include/sockets/webServer.h"
+#include "./include/utils/handleJson.h"
 
 #ifdef WIN32
 CRITICAL_SECTION CriticalSection;
@@ -34,49 +35,6 @@ int connectedSockets[MAX_CLIENT];
 struct product *serverProductList;
 int PRODUCT_NUMBER = 0;
 
-DWORD WINAPI webServer(){
-    int sockfd, portno;
-    char buffer[1024];
-    struct sockaddr_in serv_addr;
-    portno = 80;
-
-#ifdef WIN32
-    WSADATA info;
-    if (WSAStartup(MAKEWORD(1, 1), &info) == SOCKET_ERROR) {
-        timestamp();
-        perror("ERROR, can't start socket\n");
-    }
-#endif
-    sockfd = socket(AF_INET, SOCK_STREAM, 0);
-
-    if (sockfd < 0){
-        timestamp();
-        perror("ERROR opening socket\n");
-    }
-    serv_addr.sin_family = AF_INET;
-    serv_addr.sin_addr.s_addr = INADDR_ANY;
-    serv_addr.sin_port = htons(portno);
-    if (bind(sockfd, (struct sockaddr*)&serv_addr,sizeof(serv_addr)) < 0){
-        perror("\nERROR on binding");
-        return 0;
-    }
-    for(;;){
-    memset(&serv_addr, 0, sizeof(serv_addr));
-    int newsockfd = acceptNewConnection(sockfd);
-    memset(&buffer, 0, sizeof(buffer));
-    recv(newsockfd,buffer,1024-1,0);
-
-    memset(&buffer, 0, sizeof(buffer));
-    char *a = getProductJson();
-    sprintf(buffer,"\"HTTP/1.1 200 OK\r\nContent-Length: 1024\r\nContent-Type: text/html\r\n\r\n<!DOCTYPE html><script> function autoRefresh() { window.location = window.location.href; } setInterval('autoRefresh()', 2000); </script>%s", a);
-    sendToClient(newsockfd, buffer);
-    free(a);
-    a = NULL;
-    closesocket(newsockfd);
-    }
-    closesocket(sockfd);
-}
-
 int main(int argc, char* argv[]){
     if (argc < 2){
         timestamp();
@@ -87,7 +45,7 @@ int main(int argc, char* argv[]){
     readProductsFromFile("./products.csv");
     
     timestamp();
-    printf("Starting sever on port %s", argv[1]);
+    printf("Starting server on port %s", argv[1]);
     
     int sockfd, portno;
     struct sockaddr_in serv_addr;
@@ -111,11 +69,11 @@ int main(int argc, char* argv[]){
 
     #ifdef WIN32
     CreateThread(NULL, 0, webServer, NULL, 0, NULL);
-
     CreateThread(NULL, 0, handleUpdateClients, NULL, 0, NULL);
     #else
     pthread_t thread_id = 0;
-    pthread_create(&thread_id,NULL,handleUpdateClients, NULL);
+    pthread_create(&thread_id, NULL, webServer, NULL);
+    pthread_create(&thread_id, NULL, handleUpdateClients, NULL);
     #endif
 
     while(1) {
